Improve startup ordering

+++ src/node/worker.nix
@@ -22,15 +22,37 @@ in
 
     systemd.services.cfssl.enable = false; # checkme: nixos/kubernetes should disable this on worker
 
-    systemd.services.kube-certmgr-bootstrap.script = mkForce ''
-      test -f ${top.secretsPath}/ca.pem && chmod u+w ${top.secretsPath}/ca.pem
-      cp -upd /run/keys/cfssl-ca        ${top.secretsPath}/ca.pem
-      chown root:root ${top.secretsPath}/ca.pem && chmod 0444 ${top.secretsPath}/ca.pem
+    systemd.services.certmgr = {
+      preStart = mkBefore ''
+          if test ! -f /run/keys/cfssl-ca && test ! -f ${top.secretsPath}/ca.pem; then
+            exit 1
+          fi
 
-      test -f ${top.secretsPath}/apitoken.secret && chmod u+w ${top.secretsPath}/apitoken.secret
-      cp -upd /run/keys/cfssl-api-token ${top.secretsPath}/apitoken.secret
-      chown root:root ${top.secretsPath}/apitoken.secret && chmod 0400 ${top.secretsPath}/apitoken.secret
-    '';
+          if test ! -f /run/keys/cfssl-api-token && test ! -f ${top.secretsPath}/apitoken.secret; then
+            exit 1
+          fi
+
+          set +e
+
+          mkdir -p ${top.secretsPath} && chmod 0755 ${top.secretsPath}
+
+          test -f ${top.secretsPath}/ca.pem && chmod u+w ${top.secretsPath}/ca.pem
+          cp -upd /run/keys/cfssl-ca        ${top.secretsPath}/ca.pem
+          chown root:root ${top.secretsPath}/ca.pem && chmod 0444 ${top.secretsPath}/ca.pem
+
+          test -f ${top.secretsPath}/apitoken.secret && chmod u+w ${top.secretsPath}/apitoken.secret
+          cp -upd /run/keys/cfssl-api-token ${top.secretsPath}/apitoken.secret
+          chown root:root ${top.secretsPath}/apitoken.secret && chmod 0400 ${top.secretsPath}/apitoken.secret
+
+          set -e
+      '';
+    };
+
+    systemd.services.kubelet = {
+      serviceConfig = {
+        StartLimitInterval = mkForce 0;
+      };
+    };
 
     services.kubernetes = {
         roles = [ "node" ];
